+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [0.0.9]
 ### Added
 - Added `Bytes::hex()` which is an alias of `Bytes::toString()`
+- Added `Bytes.decode()`
 - Added `SExpAsAtom` type and `SExpAsPair` type.
 - Added hex string support to `b(<str>)`. In `b("0xaa", "hex")`, `0xaa` is now parsed as hex string.
 ### Fixed