+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [0.0.9]
+### Added
+- Added `Bytes::hex()` which is an alias of `Bytes::toString()`
+- Added `Bytes.decode()`
+- Added `Bytes.startswith()` and `Bytes.endswith()`
+- Added `SExpAsAtom` type and `SExpAsPair` type.
+- Added hex string support to `b(<str>)`. In `b("0xaa", "hex")`, `0xaa` is now parsed as hex string.
+### Fixed
+- Fixed value of `SExp.TRUE` and `SExp.FALSE`
+
 ## [0.0.8]
 ### Added
 - Added `b(<str>)` helper function which equals to `Bytes.from(<str>, "utf8")`
@@ -49,6 +59,7 @@
 Initial (beta) release.
 
 <!--[Unreleased]: https://github.com/Chia-Mine/clvm-js/compare/v0.0.1...v0.0.2-->
+[0.0.9]: https://github.com/Chia-Mine/clvm-js/compare/v0.0.8...v0.0.9
 [0.0.8]: https://github.com/Chia-Mine/clvm-js/compare/v0.0.7...v0.0.8
 [0.0.7]: https://github.com/Chia-Mine/clvm-js/compare/v0.0.6...v0.0.7
 [0.0.6]: https://github.com/Chia-Mine/clvm-js/compare/v0.0.5...v0.0.6